finishApp

+++ app/src/main/java/com/example/jobs/model/UserInJob.java
@@ -9,7 +9,7 @@ public class UserInJob {
     private Object email_verified_at;
     private String phone;
     private String user_type;
-    private int rating;
+    private double rating;
     private String about_me;
     private Object lat;
     private Object lng;
@@ -47,7 +47,7 @@ public class UserInJob {
         return user_type;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
